+++ NEWS.md
@@ -15,6 +15,7 @@ Fixes:
 * Fixed existing reference formatting.
 * Fixed references for missing-value implementation mockup in vignette.
 * Fixed a bug in `normalise` that resulted in relations having duplicate keys.
+* Fixed a bug in `normalise`, that resulted in schemas that didn't reproduce the given functional dependencies.
 
 # autodb 1.0.0